fixed empty directory runs

+++ examples/teslausbarchive.sh
@@ -16,52 +16,56 @@
 # 4) archive videos in local directory (copy dir to archive sub folder)
 # 5) remove archive videos in local directory after 30 days unless file .preserve is present
 #============================================================================
+shopt -s extglob # Needed for directory except loop command below
 
 destroot="/mnt2/tesla_dashcam"
 sourceroot="/mnt2/tesla"
 
 echo "Looping through $sourceroot/ for teslacam video to move"
-shopt -s extglob # Needed for command below)
 for dirsource in ${sourceroot}/!(archive|processing)/; do
-	sourcedirdate=$(basename $dirsource | cut -d'_' -f1)
-	dir_time=$(date -d $sourcedirdate +%s)
-	if [ $dir_time -le $(date -d 'now - 1 day' +%s) ] || [ $sourcedirdate == $(date -d 'now - 1 day' +%Y-%m-%d) ]; then
-		echo "$dirsource is over 24 hours old"
-		destpath="${destroot}/${sourcedirdate}"
-		if [ ! -d "$destpath" ]; then
-			echo "Creating $destpath"
-			mkdir $destpath;
+	if [ -d ${dirsource} ]; then
+		sourcedirdate=$(basename $dirsource | cut -d'_' -f1)
+		dir_time=$(date -d $sourcedirdate +%s)
+		if [ $dir_time -le $(date -d 'now - 1 day' +%s) ] || [ $sourcedirdate == $(date -d 'now - 1 day' +%Y-%m-%d) ]; then
+			echo "$dirsource is over 24 hours old"
+			destpath="${destroot}/${sourcedirdate}"
+			if [ ! -d "$destpath" ]; then
+				echo "Creating $destpath"
+				mkdir $destpath;
+			fi
+			processpath="${sourceroot}/processing/${sourcedirdate}"
+			if [ ! -d "$processpath" ]; then
+				echo "Creating $processpath"
+				mkdir $processpath;
+			fi
+			echo "Moving $disource to processing directory $processpath"
+			mv $dirsource $processpath
 		fi
-		processpath="${sourceroot}/processing/${sourcedirdate}"
-		if [ ! -d "$processpath" ]; then
-			echo "Creating $processpath"
-			mkdir $processpath;
-		fi
-		echo "Moving $disource to processing directory $processpath"
-		mv $dirsource $processpath
 	fi
 done
 
 echo "Converting Videos from processing folder with tesla_dashcam container"
 for dirprocess in ${sourceroot}/processing/*/; do
-	processdirdate=$(basename $dirprocess)
-	docker run --rm \
-	  -e TZ=America/New_York \
-	  -v $destroot/$processdirdate:/root/Videos \
-	  -v $dirprocess:/root/Import \
-	  magicalyak/tesla_dashcam \
-	  --motion_only \
-	  --merge /root/Import \
-	  --output /root/Videos/$processdirdate.mp4
-	echo "Archiving folder in $sourceroot/archive/$processdirdate"
-	mv $dirprocess $sourceroot/archive/$processdirdate
-	echo "Removing intermediate files in $destroot/$processdirdate"
-	for file in ${destroot}/${processdirdate}/*; do
-		if [[ $file == *"${processdirdate}T"* ]]; then
-			echo "Deleting $file"
-			rm -f $file
-		fi
-	done
+	if [ -d ${dirprocess} ]; then
+		processdirdate=$(basename $dirprocess)
+		docker run --rm \
+		-e TZ=America/New_York \
+		-v $destroot/$processdirdate:/root/Videos \
+		-v $dirprocess:/root/Import \
+		magicalyak/tesla_dashcam \
+		--motion_only \
+		--merge /root/Import \
+		--output /root/Videos/$processdirdate.mp4
+		echo "Archiving folder in $sourceroot/archive/$processdirdate"
+		mv $dirprocess $sourceroot/archive/$processdirdate
+		echo "Removing intermediate files in $destroot/$processdirdate"
+		for file in ${destroot}/${processdirdate}/*; do
+			if [[ $file == *"${processdirdate}T"* ]]; then
+				echo "Deleting $file"
+				rm -f $file
+			fi
+		done
+	fi
 done
